move addXmcMlirPasses into a separate file
