+++ src/Compiler/CMakeLists.txt
@@ -103,6 +103,7 @@ endif()
 
 add_onnx_mlir_library(OMOnnxToMlirPasses
   OnnxToMlirPasses.cpp
+  XmcMlirPasses.cpp
   
   EXCLUDE_FROM_OM_LIBS
